refactor: DashboardDisplayData
Changed dashboard display data to its own domain class, so we can fetch it completely independently of dashboard. This _does_ mean that editing display data does not count as an update to the dashboard as far as GORM is concerned. However we should NOT be reflecting frontend ordering changes in any backend logic anyway, so this is fine--I think

ERM-2971