ERM-2971: Back end changes to support storing location/dimensions of widgets (#93)
* feat: Dashboard display data

Added a String field to house generic display data about the dashboard. Have not explicitly modelled as this is display only data and we should not particularly care

refs ERM-2971, ERM-1795

* refactor: DashboardDisplayData

Changed dashboard display data to its own domain class, so we can fetch it completely independently of dashboard. This _does_ mean that editing display data does not count as an update to the dashboard as far as GORM is concerned. However we should NOT be reflecting frontend ordering changes in any backend logic anyway, so this is fine--I think

ERM-2971

* chore: Removed old displayData linked to dashboard object itself