feat: Reset controls
Added resetDefinitionCache and resetImplementorCache commands to ActionController to reset caches for widget definitions.

This is important post SI-34 fixes in grails-okapi, since the cache could spin up whilst a particular module is down, and need bumping manually once that module is back. This is a temporary fix, but provides a workflow for now.

Also implemented a null safety fix and fixed a logging typo

SI-34