fix: Permissions oversight (tangentially refs SI-43)
Fixes permissions oversight where a user can have permissions to view number generators but not the refdata necessary to properly display those. Becomes more of an issue upstream as you cannot then select from the refdata for editing/creating. This fix adds the refdata.read permission as a subperm to the base numberGenerator.view permission

Tangentially refs SI-43 (in that it is about permission cleanup)