fix: invertChecksum (#132)
Previously invertChecksum was used to allow the implementation to "invert" the built in checksum library's behaviour to do (base-calculation). Some checksums require instead the calculation outright so for now we re-invert to return the original checksum calculation (an inelegant solution). However while it works for numbers up to and including the base:

`10 - (10-9) = 9` etc. It falls down at the base itself `10 - (10-10) = 10`. This is because an extra step is needed to ensure that the answer itself is modulo <base>.